Adjustments
Adjusted how we read the file in the window parser so that it reads it in a stream.

This will save a ton of memory, and is really what we should have done in Part A in the first place.
This should allow us to read through the files without nearly as much memory problems.